[FIX] of the issue #17394 and #18065 (problems with GPR predictions on multitarget data)
Initially, when a multitarget data was given, the _gpr.py was not able to undo the normalisation of y_cov and y_var in the predict() function (lines 355 and 381). The problem was the fact that the code was trying to multiply two horizontal vectors in both cases and returned error "ValueError: operands could not be broadcast together with shapes (a,) (b,)" (e.g. number of samples 'a' = 4, number of target components 'b' = 3). I solved this by reshaping the y_var vector and rows of y_cov into vertical form before multiplication with the train std vector. As a result, we obtain y_std of form (n_samples, n_targetsy_cov